fix: alterado mapeamento de UserId em todas classes e feito novo migration

+++ ClinicasCRM.Core/Models/Procedure/ProcedurePackProcedure.cs
@@ -3,8 +3,8 @@
     public class ProcedurePackProcedure
     {
         public ProcedurePack ProcedurePack { get; set; } = new();
-        public int ProcedurePackId { get; set; }
+        public long ProcedurePackId { get; set; }
         public Procedure Procedure { get; set; } = new();
-        public int ProcedureId { get; set; }
+        public long ProcedureId { get; set; }
     }
 }
